+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.4.0
+
+- Modifies signature of `Inspecto.Schema.inspect/1` to return a more standard `:ok` or `:error` tuple.
+- adds support for tracking table prefix
+- adds support for tracking virtual fields
+- updates dependencies to latest
+
 ## v0.3.0
 
 Adds support for an `:h` option to specify heading level in generated HTML